+++ src/hypertoolz/core/__init__.py
@@ -1,6 +1,6 @@
 from .config import TunerConfig
+from .hypertypes import OptunaParam, ParseResult, ValidationMode
 from .results import OptimizationResults
-from .types import OptunaParam, ParseResult, ValidationMode
 
 __all__ = [
     "TunerConfig",
